requirements.txt
aiogram==3.13.1
peewee==3.17.6
python-dotenv==1.0.1
peewee-migrate==1.12.2
